Restructure components and page
