+++ packages/playground/__tests__/SvgImport.test.tsx
@@ -1,4 +1,4 @@
-import Logo from '~/assets/ink-logo-on-dark.svg';
+import { Logo } from '~/symbols';
 
 test('Given the Logo component is imported', () => {
   // When it is transformed by jest
