+++ packages/playground/__tests__/Welcome.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import { Welcome } from '~/components';
+import { Welcome } from '~/page/WasmTest/Welcome';
 
 test('Given the Welcome component is imported', () => {
   // When it is rendered
