--- packages/playground/src/components/Buttons/index.tsx
@@ -1,1 +1,0 @@
-export { ButtonWithIcon } from './ButtonWithIcon';
