+++ packages/playground/src/components/index.tsx
@@ -1,5 +1,1 @@
-export { Console } from './Console';
-export { Editor } from './Editor';
-export { Layout } from './Layout';
-export { Header } from './Header';
-export { Welcome } from './WasmTest/Welcome';
+export { ButtonWithIcon } from './ButtonWithIcon';
