+++ packages/playground/src/symbols.tsx
@@ -1,0 +1,19 @@
+import CompileIcon from '~/assets/compile.svg';
+import DownloadIcon from '~/assets/download.svg';
+import GithubRepoIcon from '~/assets/branch.svg';
+import SettingsIcon from '~/assets/settings.svg';
+import DeployIcon from '~/assets/deploy.svg';
+import DocsIcon from '~/assets/docs.svg';
+import GistIcon from '~/assets/gist.svg';
+import Logo from '~/assets/ink-logo-on-dark.svg';
+
+export {
+  CompileIcon,
+  DownloadIcon,
+  GithubRepoIcon,
+  SettingsIcon,
+  DeployIcon,
+  DocsIcon,
+  GistIcon,
+  Logo,
+};
