+++ packages/playground/src/typings/svg.d.ts
